Move from trunk/, make composer friendly
