+++ languages/table-of-contents-plus-de_DE.mo
