+++ languages/table-of-contents-plus-fr_FR.mo
