+++ languages/table-of-contents-plus-sk_SK.po
